🔨 파일명 변경 및 DiscussionInputs, SearchBook 적용 - #41
